rename file names to kebab-case (responding https://markdown-all-in-one.github.io/docs/contributing/code-style/file.html)
