+++ src/editor-context-service/context-service-in-list.ts
@@ -1,10 +1,9 @@
 'use strict'
 
 import { ExtensionContext, Position, TextDocument, window } from 'vscode';
-import { AbsContextService } from "./IContextService";
-import { isInFencedCodeBlock } from "../util/contextCheck";
+import { AbsContextService } from "./i-context-service";
 
-export class ContextServiceEditorInFencedCodeBlock extends AbsContextService {
+export class ContextServiceEditorInList extends AbsContextService {
     public onActivate(context: ExtensionContext) {
         // set initial state of context
         this.setState(false);
@@ -21,7 +20,10 @@ export class ContextServiceEditorInFencedCodeBlock extends AbsContextService {
     }
 
     private updateContextState(document: TextDocument, cursorPos: Position) {
-        if (isInFencedCodeBlock(document, cursorPos.line)) {
+        let lineText = document.lineAt(cursorPos.line).text;
+
+        let inList = /^\s*([-+*]|[0-9]+[.)]) +(\[[ x]\] +)?/.test(lineText);
+        if (inList) {
             this.setState(true);
         }
         else {
